Fixed bug where empty values could be saved
Users were able to click 'Save' in the options page before a
shoe size selection was made. Issue #2 needs to be fixed in
order to disable YASB from working when an empty value is
stored to #size, as is still the case even if the user isn't
allowed to save it that way.

[Fixes #1]