Added option to compile a free or paid version
Changing the value of the variable 'client' to 'consumer' or 'personal'
in background.js either enables or disables YASB to require an activation
key. When enabled, the software will require an activation to function,
allowing the option to sell YASB. These modifications need to be done
before compiling source.

Some code has been rearranged for clarity.

[Closes #3]